requirements.txt
# Multi-Strategy Options Scanner - Python Dependencies
# Generated: 2025-11-15

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23  # Optional ORM

# API & HTTP
requests==2.31.0
requests-ratelimiter==0.4.0

# Data Processing
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4

# Options Calculations
py-vollib==1.0.1

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.11.0

# Utilities
python-dateutil==2.8.2

# Production Server (for deployment)
gunicorn==21.2.0

# Optional: Rate Limiting
Flask-Limiter==3.5.0

# Optional: Caching
Flask-Caching==2.1.0
